Move sticky section out of the section metadata
