+++ libs/blocks/section-metadata/section-metadata.js
@@ -1,5 +1,3 @@
-import { createTag } from '../../utils/utils.js';
-
 export function handleFocalpoint(pic, child, removeChild) {
   const image = pic.querySelector('img');
   if (!child || !image) return;
@@ -17,6 +15,7 @@ export function handleFocalpoint(pic, child, removeChild) {
   const [x, y = ''] = directions;
   image.style.objectPosition = `${x} ${y}`;
 }
+
 function handleBackground(div, section) {
   const pic = div.background.content?.querySelector('picture');
   if (pic) {
@@ -32,65 +31,14 @@ function handleBackground(div, section) {
   }
 }
 
-function handleTopHeight(section) {
-  const headerHeight = document.querySelector('header').offsetHeight;
-  section.style.top = `${headerHeight}px`;
-}
-
-function promoIntersectObserve(el, stickySectionEl, options = {}) {
-  const io = new IntersectionObserver((entries, observer) => {
-    entries.forEach((entry) => {
-      if (el.classList.contains('close-sticky-section')) {
-        observer.unobserve(entry.target);
-        return;
-      }
-      const isPromoStart = entry.target === stickySectionEl;
-      const abovePromoStart = (isPromoStart && entry.isIntersecting)
-                        || stickySectionEl?.getBoundingClientRect().y > 0;
-      if (entry.isIntersecting || abovePromoStart) el.classList.add('hide-sticky-section');
-      else el.classList.remove('hide-sticky-section');
-    });
-  }, options);
-  return io;
-}
-
-function handleStickyPromobar(section) {
-  const main = document.querySelector('main');
-  section.classList.add('hide-sticky-section');
-  let stickySectionEl = null;
-  if (main.children[0] !== section) {
-    stickySectionEl = createTag('div', { class: 'section show-sticky-section' });
-    main.insertBefore(stickySectionEl, section);
-  }
-  const io = promoIntersectObserve(section, stickySectionEl);
-  if (stickySectionEl) io.observe(stickySectionEl);
-  io.observe(document.querySelector('footer'));
-}
-
-async function handleStickySection(sticky, section) {
-  const main = document.querySelector('main');
-  switch (sticky) {
-    case 'sticky-top':
-    {
-      const { debounce } = await import('../../utils/action.js');
-      window.addEventListener('resize', debounce(() => handleTopHeight(section)));
-      main.prepend(section);
-      break;
-    }
-    case 'sticky-bottom':
-      if (section.querySelector('.promobar')) handleStickyPromobar(section);
-      main.append(section);
-      break;
-    default:
-      break;
-  }
-}
-
 export async function handleStyle(text, section) {
   if (!text || !section) return;
   const styles = text.split(', ').map((style) => style.replaceAll(' ', '-'));
   const sticky = styles.find((style) => style === 'sticky-top' || style === 'sticky-bottom');
-  if (sticky) await handleStickySection(sticky, section);
+  if (sticky) {
+    const { default: handleStickySection } = await import('./sticky-section.js');
+    await handleStickySection(sticky, section);
+  }
   if (styles.includes('masonry')) styles.push('masonry-up');
   section.classList.add(...styles);
 }
