+++ test/blocks/section-metadata/mocks/body.html
@@ -64,18 +64,6 @@
       </div>
     </div>
   </div>
-  <div class="section sticky-bottom">
-    <div class="section-metadata">
-      <div>
-        <div>background</div>
-        <div>#ababab</div>
-      </div>
-      <div>
-        <div>style</div>
-        <div>sticky-bottom</div>
-      </div>
-    </div>
-  </div>
   <div class="section sticky-top">
     <div class="section-metadata">
       <div>
